+++ asynctor/contrib/fastapi.py
@@ -2,17 +2,21 @@ from __future__ import annotations
 
 import functools
 import logging
+import os
+import platform
+import re
 import sys
 from collections.abc import AsyncGenerator, Callable
 from contextlib import asynccontextmanager
 from pathlib import Path
-from typing import Annotated, Any
+from typing import Annotated, Any, TypeAlias
 
 import uvicorn
 from fastapi import Depends, FastAPI, Request
 from fastapi.routing import _merge_lifespan_context
 
 from ..client import AsyncRedis
+from ..utils import Shell, load_bool
 
 
 def register_aioredis(
@@ -110,6 +114,17 @@ def config_access_log_to_show_time(log: str = "uvicorn.access") -> None:
     logging.getLogger(log).addHandler(handler)
 
 
+PreStartFunc: TypeAlias = Callable[
+    [
+        Annotated[str, "host"],
+        Annotated[int | None, "port"],
+        Annotated[bool, "reload"],
+        Annotated[dict[str, str] | None, "docs_params"],
+    ],
+    Any,
+]
+
+
 class RunServer:
     @staticmethod
     def uvicorn_run(app: FastAPI, host: str, port: int | None, reload: bool, **kw) -> None:
@@ -151,6 +166,10 @@ class RunServer:
                 if verbose:
                     echo(f"Load `PORT = {p}` from {config_file}")
                 return int(p)
+            elif (p := getattr(gunicorn_config, "bind", "").split(":")[-1]).isdigit():
+                if verbose:
+                    echo(f"Load `bind = xxx:{p}` from {config_file}")
+                return int(p)
             elif verbose:
                 echo(f"{config_file} does not have 'PORT' attribute")
         elif verbose:
@@ -164,11 +183,14 @@ class RunServer:
         port: int | None,
         docs_params: dict | None = None,
         echo: Callable | None = None,
-    ) -> None:
-        from asynctor.utils import get_machine_ip
-
+    ) -> str:
         if host == "0.0.0.0":
-            host = get_machine_ip()
+            if declared_host := os.getenv("ASYNCTOR_HOST"):
+                host = declared_host
+            else:
+                from asynctor.utils import get_machine_ip
+
+                host = get_machine_ip()
         url = f"http://{host}:{port or 8000}{app.docs_url}"
         if docs_params:
             url += "?" + "&".join(f"{k}={v}" for k, v in docs_params.items())
@@ -181,6 +203,16 @@ class RunServer:
                 echo(url, bold=True)
             except TypeError:
                 echo(url)
+        return url
+
+    @staticmethod
+    def load_port_from_env() -> int | None:
+        if p := os.getenv("ASYNCTOR_PORT"):
+            try:
+                return int(p)
+            except ValueError:
+                ...
+        return None
 
     @classmethod
     def run(
@@ -194,6 +226,8 @@ class RunServer:
         verbose: bool,
         echo: Callable,
         docs_params: dict[str, str] | None = None,
+        pre_start: PreStartFunc | None = None,
+        open_browser: bool | None = None,
         **kw,
     ) -> None:
         if addrport:
@@ -217,7 +251,23 @@ class RunServer:
             else:
                 if verbose:
                     echo(f"Deployment dir: {deployment_dir.name!r} not found")
-        cls.echo_docs_url(app, host, port, docs_params, echo)
+        cls.echo_and_run(app, host, port, reload, docs_params, pre_start, echo, **kw)
+
+    @classmethod
+    def echo_and_run(cls, app, host, port, reload, docs_params, pre_start, echo=None, **kw) -> None:
+        if not port:
+            port = cls.load_port_from_env()
+        url = cls.echo_docs_url(app, host, port, docs_params, echo)
+        if pre_start is not None:
+            try:
+                pre_start(host=host, port=port, reload=reload, docs_params=docs_params)
+            except TypeError:
+                pre_start()
+        if kw.pop("open_browser", False) or load_bool("ASYNCTOR_BROWSER"):
+            command = "explorer" if platform.system() == "Windows" else "open"
+            if host == "0.0.0.0" and (m := re.search(r"://(.*?)[:/]", url)):
+                url = url.replace(m.group(1), "127.0.0.1")
+            Shell([command, url]).run(verbose=True)
         cls.uvicorn_run(app, host, port, reload, **kw)
 
 
@@ -229,12 +279,13 @@ def runserver(
     reload: bool = False,
     verbose: bool = False,
     docs_params: dict[str, str] | None = None,
+    pre_start: PreStartFunc | None = None,
+    open_browser: bool | None = None,
     **kw,
 ) -> None:
+    kw.update(docs_params=docs_params, pre_start=pre_start, open_browser=open_browser)
     if not (args := sys.argv[1:]):
-        RunServer.echo_docs_url(app, host, port, docs_params)
-        return RunServer.uvicorn_run(app, host, port, reload, **kw)
-
+        return RunServer.echo_and_run(app, host, port, reload, **kw)
     try:
         import typer
     except ImportError: